refactor: split up COSE stuff a bit more
