+++ src/lib.rs
@@ -314,7 +314,6 @@
     clippy::type_complexity
 )]
 #![cfg_attr(not(feature = "std"), no_std)]
-#![cfg_attr(feature = "nightly", feature(min_specialization))]
 #[macro_use]
 extern crate alloc;
 #[macro_use]
