+++ src/token/cose/sign/tests.rs
@@ -1,227 +1,24 @@
 use crate::token::cose::crypto_impl::openssl::OpensslContext;
-use crate::token::cose::sign::{CoseSign1Ext, CoseSignBuilderExt, CoseSignExt};
-use crate::token::CoseSign1BuilderExt;
+use crate::token::cose::sign::CoseSign1BuilderExt;
+use crate::token::cose::sign::CoseSign1Ext;
+use crate::token::cose::sign::{CoseSignBuilderExt, CoseSignExt};
+use crate::token::cose::test_helper::{
+    TestCase, TestCaseFailures, TestCaseInput, TestCaseSign, TestCaseSigner,
+};
 use crate::CoseSignCipher;
-use base64::engine::general_purpose::URL_SAFE_NO_PAD;
 use base64::Engine;
-use core::fmt::Formatter;
-use coset::iana::{Algorithm, EnumI64};
+use coset::iana::EnumI64;
 use coset::{
-    iana, AsCborValue, CborSerializable, CoseError, CoseKey, CoseKeyBuilder, CoseSign, CoseSign1,
-    CoseSign1Builder, CoseSignBuilder, CoseSignature, CoseSignatureBuilder, Header, HeaderBuilder,
-    Label, TaggedCborSerializable,
+    AsCborValue, CborSerializable, CoseError, CoseKey, CoseSign, CoseSign1, CoseSign1Builder,
+    CoseSignBuilder, CoseSignature, CoseSignatureBuilder, Header, Label, TaggedCborSerializable,
 };
 use hex::FromHex;
-use openssl::sign::Signer;
 use rstest::rstest;
 use serde::de::{MapAccess, Visitor};
-use serde::{de, Deserialize, Deserializer};
+use serde::{Deserialize, Deserializer};
 use serde_json::Value;
 use std::any::Any;
-use std::collections::{HashMap, HashSet};
 use std::path::PathBuf;
-
-fn deserialize_key<'de, D>(deserializer: D) -> Result<CoseKey, D::Error>
-where
-    D: Deserializer<'de>,
-{
-    let key_obj = serde_json::Map::deserialize(deserializer)?;
-    match key_obj.get("kty").map(Value::as_str).flatten() {
-        Some("EC") => {
-            let curve = match key_obj.get("crv").map(Value::as_str).flatten() {
-                Some("P-256") => iana::EllipticCurve::P_256,
-                Some("P-384") => iana::EllipticCurve::P_384,
-                Some("P-521") => iana::EllipticCurve::P_521,
-                _ => return Err(de::Error::custom("COSE key does not have valid curve")),
-            };
-
-            let x = if let Some(v) = key_obj
-                .get("x")
-                .map(Value::as_str)
-                .flatten()
-                .map(|v| URL_SAFE_NO_PAD.decode(v).ok())
-                .flatten()
-            {
-                v
-            } else {
-                return Err(de::Error::custom("COSE key does not have valid x"));
-            };
-
-            let y = if let Some(v) = key_obj
-                .get("y")
-                .map(Value::as_str)
-                .flatten()
-                .map(|v| URL_SAFE_NO_PAD.decode(v).ok())
-                .flatten()
-            {
-                v
-            } else {
-                return Err(de::Error::custom("COSE key does not have valid x"));
-            };
-            let d = if let Some(v) = key_obj
-                .get("d")
-                .map(Value::as_str)
-                .flatten()
-                .map(|v| URL_SAFE_NO_PAD.decode(v).ok())
-                .flatten()
-            {
-                v
-            } else {
-                return Err(de::Error::custom("COSE key does not have valid x"));
-            };
-
-            let mut builder = CoseKeyBuilder::new_ec2_priv_key(curve, x, y, d);
-
-            if let Some(v) = key_obj.get("kid").map(Value::as_str).flatten() {
-                builder = builder.key_id(v.to_string().into_bytes())
-            }
-
-            Ok(builder.build())
-        }
-        _ => Err(de::Error::custom("COSE key does not have valid key type")),
-    }
-}
-
-fn deserialize_header<'de, D>(deserializer: D) -> Result<Option<Header>, D::Error>
-where
-    D: Deserializer<'de>,
-{
-    let hdr_obj =
-        if let Some(hdr) = Option::<serde_json::Map<String, Value>>::deserialize(deserializer)? {
-            hdr
-        } else {
-            return Ok(None);
-        };
-
-    let mut builder = HeaderBuilder::new();
-
-    if let Some(v) = hdr_obj.get("kid").map(Value::as_str).flatten() {
-        builder = builder.key_id(v.to_string().into_bytes());
-    }
-    if let Some(v) = hdr_obj
-        .get("kid_hex")
-        .map(Value::as_str)
-        .flatten()
-        .map(hex::decode)
-    {
-        builder = builder
-            .key_id(v.map_err(|e| de::Error::custom("could not parse test case key ID hex"))?);
-    }
-
-    builder = match hdr_obj.get("alg").map(Value::as_str).flatten() {
-        Some("ES256") => builder.algorithm(Algorithm::ES256),
-        Some("ES384") => builder.algorithm(Algorithm::ES384),
-        Some("ES512") => builder.algorithm(Algorithm::ES512),
-        Some(_) => return Err(de::Error::custom("could not parse test case algorithm")),
-        None => builder,
-    };
-
-    builder = match hdr_obj.get("ctyp").map(Value::as_i64).flatten() {
-        Some(v) => {
-            let content_format = iana::CoapContentFormat::from_i64(v)
-                .ok_or(de::Error::custom("could not parse test case algorithm"))?;
-            builder.content_format(content_format)
-        }
-        None => builder,
-    };
-
-    Ok(Some(builder.build()))
-}
-
-fn deserialize_algorithm<'de, D>(deserializer: D) -> Result<Option<Algorithm>, D::Error>
-where
-    D: Deserializer<'de>,
-{
-    let alg = if let Some(alg) = Option::<Value>::deserialize(deserializer)? {
-        alg
-    } else {
-        return Ok(None);
-    };
-    match alg.as_str() {
-        Some("ES256") => Ok(Some(Algorithm::ES256)),
-        Some("ES384") => Ok(Some(Algorithm::ES384)),
-        Some("ES512") => Ok(Some(Algorithm::ES512)),
-        _ => Err(de::Error::custom("could not parse test case algorithm")),
-    }
-}
-
-#[derive(Deserialize, Debug, Clone)]
-pub struct TestCase {
-    title: String,
-    description: Option<String>,
-    #[serde(default)]
-    fail: bool,
-    input: TestCaseInput,
-    intermediates: Option<TestCaseIntermediates>,
-    output: TestCaseOutput,
-}
-
-#[derive(Deserialize, Debug, Clone, Default)]
-pub struct TestCaseFailures {
-    #[serde(rename = "ChangeTag")]
-    change_tag: Option<u64>,
-    #[serde(rename = "ChangeCBORTag")]
-    change_cbor_tag: Option<u64>,
-    #[serde(rename = "ChangeAttr")]
-    change_attribute: Option<HashMap<String, Value>>,
-    #[serde(
-        rename = "AddProtected",
-        deserialize_with = "deserialize_header",
-        default
-    )]
-    add_protected_headers: Option<Header>,
-    #[serde(
-        rename = "RemoveProtected",
-        deserialize_with = "deserialize_header",
-        default
-    )]
-    remove_protected_headers: Option<Header>,
-}
-
-#[derive(Deserialize, Debug, Clone)]
-pub struct TestCaseInput {
-    plaintext: String,
-    #[serde(default)]
-    detached: bool,
-    sign0: Option<TestCaseSigner>,
-    sign: Option<TestCaseSign>,
-    #[serde(default)]
-    failures: TestCaseFailures,
-}
-
-#[derive(Deserialize, Debug, Clone)]
-pub struct TestCaseSign {
-    #[serde(deserialize_with = "deserialize_header", default)]
-    unprotected: Option<Header>,
-    #[serde(deserialize_with = "deserialize_header", default)]
-    protected: Option<Header>,
-    signers: Vec<TestCaseSigner>,
-}
-
-#[derive(Deserialize, Debug, Clone)]
-pub struct TestCaseSigner {
-    #[serde(deserialize_with = "deserialize_key")]
-    key: CoseKey,
-    #[serde(deserialize_with = "deserialize_header", default)]
-    unprotected: Option<Header>,
-    #[serde(deserialize_with = "deserialize_header", default)]
-    protected: Option<Header>,
-    #[serde(deserialize_with = "deserialize_algorithm", default)]
-    alg: Option<Algorithm>,
-    #[serde(deserialize_with = "hex::deserialize", default)]
-    external: Vec<u8>,
-    #[serde(default)]
-    failures: TestCaseFailures,
-}
-
-#[derive(Deserialize, Debug, Clone)]
-pub struct TestCaseIntermediates {}
-
-#[derive(Deserialize, Debug, Clone)]
-pub struct TestCaseOutput {
-    #[serde(deserialize_with = "hex::deserialize")]
-    cbor: Vec<u8>,
-}
 
 fn serialize_sign1_and_apply_failures(
     test_case_input: &TestCaseInput,
